Added Contact class to include common methods for Address, Email, Phone classes Removed common methods for Address, Email, Phone classes
